final-job/signer/requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
cffi==1.12.3
Click==7.0
cryptography==2.7
Flask==1.0.3
importlib-metadata==0.18
isort==4.3.20
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.0.0
packaging==19.0
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pylint==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.0
pytest==4.6.3
python-dotenv==0.10.3
six==1.12.0
typed-ast==1.4.0
wcwidth==0.1.7
Werkzeug==0.15.4
wrapt==1.11.2
zipp==0.5.1
